+++ rog-control-center/src/app.rs
@@ -28,7 +28,9 @@ pub struct RogApp<'a> {
     pub config: Config,
     pub asus_dbus: RogDbusClientBlocking<'a>,
     /// Oscillator in percentage
-    pub oscillator: Arc<AtomicU8>,
+    pub oscillator1: Arc<AtomicU8>,
+    pub oscillator2: Arc<AtomicU8>,
+    pub oscillator3: Arc<AtomicU8>,
     /// Frequency of oscillation
     pub oscillator_freq: Arc<AtomicU8>,
     /// A toggle that toggles true/false when the oscillator reaches 0
@@ -49,8 +51,14 @@ impl<'a> RogApp<'a> {
 
         // Set up an oscillator to run on a thread.
         // Helpful for visual effects like colour pulse.
-        let oscillator = Arc::new(AtomicU8::new(0));
-        let oscillator1 = oscillator.clone();
+        let oscillator1 = Arc::new(AtomicU8::new(0));
+        let oscillator2 = Arc::new(AtomicU8::new(0));
+        let oscillator3 = Arc::new(AtomicU8::new(0));
+
+        let oscillator1_1 = oscillator1.clone();
+        let oscillator1_2 = oscillator2.clone();
+        let oscillator1_3 = oscillator3.clone();
+
         let oscillator_freq = Arc::new(AtomicU8::new(5));
         let oscillator_freq1 = oscillator_freq.clone();
         let oscillator_toggle = Arc::new(AtomicBool::new(false));
@@ -62,19 +70,27 @@ impl<'a> RogApp<'a> {
                 let time = started.elapsed();
                 // 32 = slow, 16 = med, 8 = fast
                 let scale = oscillator_freq1.load(Ordering::SeqCst) as f64;
-                let elapsed = time.as_millis() as f64 / 10000.0;
-                let tmp = ((scale * elapsed * PI).cos()).abs();
-                if tmp <= 0.1 && !toggled {
+                let elapsed1 = (time.as_millis() as f64 + 333.0) / 10000.0;
+                let elapsed2 = (time.as_millis() as f64 + 666.0) / 10000.0;
+                let elapsed3 = (time.as_millis() as f64 + 999.0) / 10000.0;
+                let tmp1 = ((scale * elapsed1 * PI).cos()).abs();
+                let tmp2 = ((scale * 0.85 * elapsed2 * PI).cos()).abs();
+                let tmp3 = ((scale * 0.7 * elapsed3 * PI).cos()).abs();
+                if tmp1 <= 0.1 && !toggled {
                     let s = oscillator_toggle1.load(Ordering::SeqCst);
                     oscillator_toggle1.store(!s, Ordering::SeqCst);
                     toggled = true;
-                } else if tmp > 0.9 {
+                } else if tmp1 > 0.9 {
                     toggled = false;
                 }
 
-                let tmp = (255.0 * tmp * 100.0 / 255.0) as u8;
+                let tmp1 = (255.0 * tmp1 * 100.0 / 255.0) as u8;
+                let tmp2 = (255.0 * tmp2 * 100.0 / 255.0) as u8;
+                let tmp3 = (255.0 * tmp3 * 100.0 / 255.0) as u8;
 
-                oscillator1.store(tmp, Ordering::SeqCst);
+                oscillator1_1.store(tmp1, Ordering::SeqCst);
+                oscillator1_2.store(tmp2, Ordering::SeqCst);
+                oscillator1_3.store(tmp3, Ordering::SeqCst);
                 std::thread::sleep(Duration::from_millis(33));
             }
         });
@@ -87,7 +103,9 @@ impl<'a> RogApp<'a> {
             running_in_bg: start_closed,
             config,
             asus_dbus: dbus,
-            oscillator,
+            oscillator1,
+            oscillator2,
+            oscillator3,
             oscillator_toggle,
             oscillator_freq,
         })
@@ -168,8 +186,9 @@ impl<'a> eframe::App for RogApp<'a> {
             self.system_page(ctx);
         } else if page == Page::AuraEffects {
             self.aura_page(ctx);
-        } else if page == Page::AnimeMatrix {
-            self.anime_page(ctx);
+        // TODO: Anime page is not complete
+        // } else if page == Page::AnimeMatrix {
+        //     self.anime_page(ctx);
         } else if page == Page::FanCurves {
             self.fan_curve_page(ctx);
         }